+++ lib/src/state/state_controller.dart
@@ -1,7 +1,9 @@
-import 'enum/direction.dart';
-import 'enum/drawer_state.dart';
+import '../enum/direction.dart';
+import '../enum/drawer_state.dart';
 
 mixin StateController {
+  bool get needHeightInitialize;
+
   double get nowHeight;
 
   DrawerState get drawerState;